rough sketch of datastore-core main classes with no okio dependency.
All okio dependencies move to datastore-core-okio.  However tests take a dependency on  datastore-core-okio to run without significant rewrite.
