+++ datastore/datastore-core-okio/src/nativeMain/kotlin/androidx.datastore.core.okio/DatastoreFactoryOkio.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.datastore.core.okio
+
+import androidx.datastore.core.DataMigration
+import androidx.datastore.core.DataStore
+import androidx.datastore.core.DataStoreFactory
+import androidx.datastore.core.Serializer
+import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
+import kotlinx.coroutines.CoroutineScope
+import okio.FileSystem
+import okio.Path
+
+object DatastoreFactoryOkio {
+    fun <T> create(
+        fileSystem: FileSystem,
+        producePath: () -> Path,
+        serializer: Serializer<T>,
+        corruptionHandler: ReplaceFileCorruptionHandler<T>?,
+        migrations: List<DataMigration<T>>,
+        scope: CoroutineScope
+    ): DataStore<T> = DataStoreFactory.create(
+        storage = OkioStorage(fileSystem, producePath, serializer),
+        corruptionHandler = corruptionHandler,
+        migrations = migrations,
+        scope = scope
+    )
+}
