+++ datastore/datastore-internal-test-utils/src/nativeMain/kotlin/androidx/datastore/core/TestActuals.native.kt
@@ -16,6 +16,7 @@
 
 package androidx.datastore.core
 
+import androidx.datastore.core.okio.OkioStorage
 import kotlin.random.Random
 import okio.BufferedSink
 import okio.BufferedSource
